# configs/swagger.json
{
	"openapi": "3.0.2",
	"info": {
		"title": "AdPushup Console",
		"description": "Description",
		"version": "1.0.0",
		"contact": {
			"name": "AdPushup",
			"email": "contact@AdPushup.com",
			"url": "https://www.adpushup.com/"
		},
		"license": {
			"name": "MIT",
			"url": "https://opensource.org/licenses/MIT"
		}
	},
	"servers": [
		{
			"url": "/api/",
			"description": "Development Server"
		}
	],
	"tags": [
		{
			"name": "indexController"
		}
	],
	"paths": {
		"/login": {
			"post": {
				"summary": "Set JWT Token in cookie",
				"tags": ["indexController"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string",
										"format": "password"
									}
								},
								"required": ["email", "password"]
							}
						}
					}
				},
				"security": [],
				"responses": {
					"200": {
						"description": "ok",
						"headers": {
							"Set-Cookie": {
								"schema": {
									"type": "string",
									"example": "authToken=abcde12345; Path=/; HttpOnly"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"401": {
						"description": "unauthorized",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"400": {
						"description": "BAD_REQUEST",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/signup": {
			"post": {
				"summary": "Create new user and set jwt in cookie",
				"tags": ["indexController"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"name": {
										"type": "string"
									},
									"email": {
										"type": "string",
										"format": "email"
									},
									"password": {
										"type": "string",
										"format": "password"
									},
									"site": {
										"type": "string",
										"format": "uri"
									},
									"websiteRevenue": {
										"type": "string",
										"enum": [
											"999",
											"1000-2500",
											"2500-5000",
											"5000-10000",
											"10000-50000",
											"50000-200000",
											"200001"
										]
									}
								},
								"required": ["name", "email", "password", "site", "websiteRevenue"]
							}
						}
					}
				},
				"security": [],
				"responses": {
					"200": {
						"description": "success",
						"headers": {
							"Set-Cookie": {
								"schema": {
									"type": "string",
									"example": "authToken=abcde12345; Path=/; HttpOnly"
								}
							}
						},
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"500": {
						"description": "internal server error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					},
					"400": {
						"description": "input validation errors",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/forgotPassword": {
			"post": {
				"summary": "forgot password",
				"tags": ["indexController"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									}
								},
								"required": ["email"]
							}
						}
					}
				},
				"security": [],
				"responses": {
					"200": {
						"description": "email sent",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"400": {
						"description": "input validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "user not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		},
		"/resetPassword": {
			"post": {
				"summary": "reset password",
				"tags": ["indexController"],
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"email": {
										"type": "string",
										"format": "email"
									},
									"key": "string",
									"password": {
										"type": "string",
										"format": "password"
									}
								},
								"required": ["email", "key", "password"]
							}
						}
					}
				},
				"security": [],
				"responses": {
					"200": {
						"description": "email sent",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/SuccessResponse"
								}
							}
						}
					},
					"400": {
						"description": "input validation error",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/FieldValidationErrorResponse"
								}
							}
						}
					},
					"404": {
						"description": "user not found",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ErrorResponse"
								}
							}
						}
					}
				}
			}
		}
	},
	"components": {
		"schemas": {
			"ErrorResponse": {
				"type": "object",
				"properties": {
					"error": {
						"type": "string"
					}
				}
			},
			"FieldValidationErrorResponse": {
				"type": "object",
				"properties": {
					"errors": {
						"type": "array",
						"items": {
							"type": "object"
						}
					}
				}
			},
			"SuccessResponse": {
				"properties": {
					"success": {
						"type": "string"
					}
				}
			}
		},
		"securitySchemes": {
			"cookieAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "user"
			}
		}
	}
}